Add cube to demonstrate depth issue
This change simply replaces the quad with a cube to demonstrate an issue
with overlapping triangles. We fix this by creating a texture that only
writes the output of a fragment for a given triangle if the depth value
is less than the current depth value for that pixel.

We initialize the depth to a value of 1.0, so the first thing that gets
drawn within the NDC, will be less than 1.0 and therefore drawn to the
screen.