Introduce entity-component abstraction
This abstraction significantly improves code reuse and flexibility. This
does not implement a traditional ECS in the data layout sense. We are
highly unlikely to run into performance issues for the scope of this
game.

The difference in the API, in this case, is simply that components can
both be bags of data and/or perform actions. It is up to the developer
to choose where to draw the lines on what each component should manage.
Generally, just keep it as little as possible to improve reuse.