Add bullet shooting
This required quite a bit of refactoring, but essentially the
calculation of the view-projection matrix has been moved into a resource
so that it is calculated once per frame. Above that, the inverse
view-projection matrix is calculated up front per frame for use by input
systems.

In this case, the mouse click is mapped to a world space coordinate
which is used for raycasting from the eye through the world space
coordinate.