Add behavior trees
This is quite the commit. I'm moving this to a separate branch as it is
more for my own learning.

For this example, I've added the ability to print the current state of a
behavior tree. The enemy just goes from "Chase" and "Shoot" to "Wander",
so it's not much more than a state machine in its current form.