Add debug client/server with behavior tree viewer
This allows the game to offload debug capabilities to a separate client.
That way, the game can be fully isolated from development tooling. I
grabbed the d3-graphviz package so that behavior trees could be emitted
as a simple graphviz string. The debug client updates this view in
real time so that you can reason about how the enemy "brain" is working.