Refactor rendering into entity-component structure
This is not its own separate post, but rather just a refactor. This
embraces the entity-component model more so that the main file is mostly
just an application runner. This emphasizes that the world doesn't
necessarily need to have a rendering component. For example, splitting
this out into a server-client architecture would be significantly easier
with this change.

The GpuResources resource is now called Renderer as it's more generally
related to rendering our 3D world. Each render pass is maintained as a
separate pipeline (basic and shadow for this game). Samplers have been
renamed to simple be the "nearest" and "linear" samplers.

For this game, the uniforms are kept constant across pipelines so they
are managed within the renderer. The bind groups also follow the same
layout, but that's just for convenience, there is no reason to keep
these similar. It does mean the bind group could be pulled up into the
renderer to avoid having to rebind for each pass, but I don't see a need
for that as-is.