Replace camera state with transform
This replaces both the camera and directional light functionality with
the original transform component. Now that the rotor is an available
construct, it's possible to model the yaw and pitch logically but apply
it directly to the transform.

I've kept the yaw and pitch as logical concepts, but you can optionally
operate on the rotor directly. It really depends on your use case.

This has effectively removed one restriction. The camera can now pitch
at any angle without issue. Previously, it was restricted to to the
range (-90, 90).